Move main.go to cr directory
